Adjust timing to account for HP's buffering everything
On the HP Z80 adapter, they buffer the address bus, the data bus,
and every other signal through a 74LS241. They then have to add
more delay to WR and IORQ to make the timing work, and for some
reason they invert WR while they're at it.

We don't have room to buffer everything, so we can't follow HP's
design without expanding the board beyond the RC-2014 form factor.
Since we aren't buffering ADDR or DATA, we can't add delays to
the other control signals, either. We just need to add the extra
delays on WR and IORQ. We keep WR inverted for compatibility.

The HP Z80 adapter also inverts IORQ and MREQ for use as the clocks
of the state analyzer. I don't know why. We can't invert them (that
would add unwanted delay), so the user will need to configure the
Master Clock for rising edges on J and K, rather than falling edges
as with the HP adapter.

This change leaves most of U1 and U2 unused. The data sheet says
we need to tie off the unused inputs. They are now tied to ground
through individual visible traces on the back side, so they can
be cut if the gates are needed for mods.