Match analyzer setup to Rev 1 board
J and K clocks (MREQ and IORQ) are no longer inverted.
This diverges from the HP definition but is necessary for
timing since we aren't buffering every signal.

STAT symbol was set up for an earlier board design. Corrected
the patterns to match STAT as defined by HP and implemented
by Rev 1 module.

Removed some low-level signals from the listing view.

Added user symbols for convenience:
BUS encodes the state of BUSRQ and BUSACK as Z80, DMA, REQ, or REL.
PAGE is displayed as RAM or ROM.

The inverse assembler now appears to be working correctly,
on a breadboarded version of the Rev 1 design.